% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-progress.R
\name{check_progress}
\alias{check_progress}
\title{Check the Progress of the Simulation Study}
\usage{
check_progress()
}
\value{
The tibble with an updated 'ran' column
}
\description{
This function loads the simulation settings from the file
'simulations/simulation-settings.rds', extracts filenames, and checks whether
the corresponding parameter settings ran by examining the existence of
associated files in the 'simulations/results/' folder. It then updates the
column 'ran'.
}
\details{
Finally, the updated tibble is written back to the original file.
}
\examples{
check_progress()

}
