% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-file-name-to-vector.R
\name{convert_file_name_to_vector}
\alias{convert_file_name_to_vector}
\title{Convert File Name to Vector}
\usage{
convert_file_name_to_vector(file_name)
}
\arguments{
\item{file_name}{A character string representing a file name with the ".rds" extension.}
}
\value{
A character vector.
}
\description{
This function takes a file name generated by create_unique_file_name_from_row
and converts it into a vector with an arbitrary number of elements.
}
\examples{
\dontrun{
# Example usage:
file_name <- "abc_123_output.rds"
vector <- convert_file_name_to_vector(file_name)
}

}
