% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-cluster.R
\name{create_cluster}
\alias{create_cluster}
\title{Create and Export Cluster for Parallel Computation}
\usage{
create_cluster(list_needed_functions_variables = list(), num_workers = 2)
}
\arguments{
\item{list_needed_functions_variables}{List with all the functions and
variables that are needed by each process (default is an empty list)}

\item{num_workers}{The number of workers in the cluster (default is 2).}
}
\value{
A parallel cluster object.
}
\description{
This function creates a parallel cluster, exports necessary functions and data,
and returns the created cluster for parallel computations.
}
\examples{
\dontrun{
# Example usage:
cluster <- create_and_export_cluster(num_workers = 2)
# Perform parallel computations using the created cluster
stopCluster(cluster)  # Don't forget to stop the cluster when done
}

}
