% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-directory-if-not-exists.R
\name{create_directory_if_not_exists}
\alias{create_directory_if_not_exists}
\title{Create a directory if it does not exist already}
\usage{
create_directory_if_not_exists(directory_path)
}
\arguments{
\item{directory_path}{Character string specifying the path of the directory to create.}
}
\value{
Invisible NULL. The function is called for its side effect of creating a directory.
}
\description{
This function checks if a specified directory exists. If the directory does
not exist, it creates the directory along with any necessary parent
directories.
}
\examples{
\dontrun{
# Example usage:
create_directory_if_not_exists("path/to/your/directory")
}

}
