% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-unique-file-name-from-row.R
\name{create_unique_file_name_from_row}
\alias{create_unique_file_name_from_row}
\title{Create Unique File Name from Tibble Row}
\usage{
create_unique_file_name_from_row(tibble_row)
}
\arguments{
\item{tibble_row}{A single row of a tibble.}
}
\value{
A character string representing a unique file name with the ".rds" extension.
}
\description{
This function creates a unique file name from a row in a tibble by concatenating
all values in the row and appending the extension ".rds".
}
\examples{
\dontrun{
# Example usage:
tibble_row <- tibble(param1 = "abc", param2 = 123, result = "output")
file_name <- create_unique_file_name_from_row(tibble_row)
}

}
