% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize-simulation-settings.R
\name{initialize_simulation_settings}
\alias{initialize_simulation_settings}
\title{Initialize Simulation Settings}
\usage{
initialize_simulation_settings(input_tibble, repetitions)
}
\arguments{
\item{input_tibble}{A tibble containing parameter settings.}

\item{repetitions}{Number of times to repeat the tibble.}
}
\value{
A new tibble with repeated rows and additional columns for tracking simulation parameters.
}
\description{
Creates simulation parameter settings by repeating a tibble with specified parameters.
}
\details{
This function initializes simulation settings by repeating a tibble of parameter settings a specified
number of times. It adds columns for replication index, whether it has been run, whether it should be ignored,
random seed, job ID, and a unique filename for each row. Before proceeding, it checks if the 'simulations'
directory exists. If not, it stops and displays an error message. It also checks if the 'simulation-settings.rds'
file already exists. If it does, the function stops and displays an error message.
}
\examples{
\dontrun{
# Assuming 'my_tibble' is your tibble with parameter settings
sim_settings <- initialize_simulation_settings(my_tibble, repetitions = 3)
}

}
