% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-simulation-settings.R
\name{load_simulation_settings}
\alias{load_simulation_settings}
\title{Load Simulation Settings from File}
\usage{
load_simulation_settings()
}
\value{
A list containing the simulation settings.
}
\description{
This function checks whether the file simulations/simulation-settings.rds exists.
If the file exists, it loads and returns its content. Otherwise, it returns an error.
}
\examples{
\dontrun{
# Example usage:
settings <- load_simulation_settings()
}

}
