% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-this.R
\name{log_this}
\alias{log_this}
\title{Log This}
\usage{
log_this(...)
}
\arguments{
\item{...}{One or more character strings representing the lines to be logged.}
}
\value{
None (invisibly returns NULL).
}
\description{
This function appends one or multiple lines to the log file.
}
\examples{
log_this("Simulation started at", Sys.time())
log_this("Simulation in progress...")
log_this("Simulation completed at", Sys.time())

}
