% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-rds-file.R
\name{modify_rds_file}
\alias{modify_rds_file}
\title{Modify RDS File}
\usage{
modify_rds_file(rds_file, action, item_name, new_value = NULL)
}
\arguments{
\item{rds_file}{Character string specifying the path to the existing RDS file.}

\item{action}{Character string specifying the action to perform ("add", "remove", or "change").}

\item{item_name}{Character string specifying the name of the item to add, remove, or change.}

\item{new_value}{The new value to add or change (ignored for "remove" action).}
}
\value{
Invisible NULL.
}
\description{
This function allows adding, removing, or changing items in an existing RDS file.
}
\examples{
\dontrun{
# Example usage:
# Add a new item
modify_rds_file("existing_file.rds", action = "add", item_name = "new_item", new_value = c(1, 2, 3))

# Remove an item
modify_rds_file("existing_file.rds", action = "remove", item_name = "item_to_remove")

# Change the value of an item
modify_rds_file("existing_file.rds", action = "change", item_name = "existing_item", new_value = "new_value")
}

}
