% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset-simulation.R
\name{reset_simulation}
\alias{reset_simulation}
\title{Reset Simulation}
\usage{
reset_simulation()
}
\value{
Prints a message indicating whether the folder and its contents were
successfully deleted or if the deletion was canceled by the user.
}
\description{
Deletes a fixed folder and its contents after user confirmation.
}
\details{
This function deletes a fixed folder and its contents. It asks the user for
confirmation before proceeding with the deletion.
}
\examples{
\dontrun{
reset_simulation()
}

}
