% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-simulation.R
\name{setup_simulation}
\alias{setup_simulation}
\title{Setup Simulation}
\usage{
setup_simulation(simulation_name)
}
\arguments{
\item{simulation_name}{Character string specifying the name of the simulation.}
}
\value{
Invisible NULL.
}
\description{
This function initializes a simulation by creating a directory for simulations
and a subdirectory for results. It also creates a "global.rds" file to store
global information about the simulation.
}
\examples{
\dontrun{
# Example usage:
setup_simulation("my_simulation")
# Run your simulation steps here
}

}
