% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-results-simulation.R
\name{process_results_simulation}
\alias{process_results_simulation}
\title{Process Simulation Results}
\usage{
process_results_simulation(cluster, process_function)
}
\arguments{
\item{cluster}{A parallel cluster object created using functions from the
parallel package, e.g., \code{simtracker::create_cluster}}

\item{process_function}{A function to be applied to each row of simulation_settings.}
}
\value{
The result of applying the specified function to each row of simulation_settings.
}
\description{
This function retrieves simulation settings using simtracker::check_progress.
Any simulations that have not run yet (checked by the 'ran' field being TRUE) are ignored,
and the user is informed by a warning. The specified function is applied to each row
of the simulation_settings tibble.
}
\examples{
\dontrun{
# Example of using process_results_simulation with a custom process function
result <- process_results_simulation(my_custom_process_function)
}

}
